Remove words of 3 or fewer letters from dictionary
If they don't count as words, we shouldn't include them.
This eliminates the awkward problem when you play "aah" in response
to "aa", and then suddenly the computer can't find a valid response,
but you didn't lose by making a word, you lost by not having a word,
and yet the table of appends said your play was valid.
This also eliminates the need for a bunch of word length checking.

Also removed advt because what's that.