Reduce computation required for appends/prepends
On startup, we examine every substring of every word in the dictionary
in order to compute the tables of appends and prepends.  The current
dictionary has 3,334,625 substrings.
650,765 of those substrings are of length 1, but are examined in order
to confirm that all 26 english letters are valid starts.
It may be worth precomputing all length 2 and 3 entries as well, as
(for appends) there are only 535 and 5448 respectively, but computing
each requires looking at ~500,000 substrings.

histogram of substring lengths = [650765, 574560, 498355, 422150, \
  345945, 272428, 204038, 143881, 95189, 58923, 33941, 18161, 9004, \
  4169, 1834, 758, 311, 120, 53, 23, 11, 5, 1]

count of actual table entries for appends = [26, 535, 5448, 22904, \
  35389, 29386, 21140, 14270, 9248, 5621, 3014, 1466, 630, 259, 98, \
  39, 19, 7, 6, 5, 3, 3, 1]