Reduce memory usage for appends/prepends
For every sequence of letters that is a substring of a dictionary word
we store a list of characters which could be appended/prepended to it.
This change replaces that list of characters with a bitset (26 options
fitting into a single int).  The cost of unpacking that bitset into a
list on each play shouldn't be noticeable.